add .local everywhere
